Break up CA interface
Moves each struct and interface into its own file
