+++ pkg/ca/csc.go
@@ -1,4 +1,4 @@
-// Copyright 2021 The Sigstore Authors.
+// Copyright 2022 The Sigstore Authors.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,18 +11,13 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//
 
 package ca
 
 import (
-	"context"
-	"crypto"
 	"crypto/x509"
 	"strings"
 
-	ct "github.com/google/certificate-transparency-go"
-	"github.com/sigstore/fulcio/pkg/challenges"
 	"github.com/sigstore/sigstore/pkg/cryptoutils"
 )
 
@@ -31,17 +26,6 @@ type CodeSigningCertificate struct {
 	FinalChain       []*x509.Certificate
 	finalPEM         string
 	finalChainPEM    []string
-}
-
-// CodeSigningPreCertificate holds a precertificate and chain.
-type CodeSigningPreCertificate struct {
-	// PreCert contains the precertificate. Not a valid certificate due to a critical poison extension.
-	PreCert *x509.Certificate
-	// CertChain contains the certificate chain to verify the precertificate.
-	CertChain []*x509.Certificate
-	// PrivateKey contains the signing key used to sign the precertificate. Will be used to sign the certificate.
-	// Included in case the signing key is rotated in between precertificate generation and final issuance.
-	PrivateKey crypto.Signer
 }
 
 func CreateCSCFromPEM(cert string, chain []string) (*CodeSigningCertificate, error) {
@@ -112,19 +96,3 @@ func (c *CodeSigningCertificate) ChainPEM() ([]string, error) {
 	}
 	return c.finalChainPEM, nil
 }
-
-// CertificateAuthority implements certificate creation with a detached SCT and fetching the CA trust bundle.
-type CertificateAuthority interface {
-	CreateCertificate(ctx context.Context, challenge *challenges.ChallengeResult) (*CodeSigningCertificate, error)
-	Root(ctx context.Context) ([]byte, error)
-}
-
-// EmbeddedSCTCA implements precertificate and certificate issuance. Certificates will contain an embedded SCT.
-type EmbeddedSCTCA interface {
-	CreatePrecertificate(ctx context.Context, challenge *challenges.ChallengeResult) (*CodeSigningPreCertificate, error)
-	IssueFinalCertificate(ctx context.Context, precert *CodeSigningPreCertificate, sct *ct.SignedCertificateTimestamp) (*CodeSigningCertificate, error)
-}
-
-// ValidationError indicates that there is an issue with the content in the HTTP Request that
-// should result in an HTTP 400 Bad Request error being returned to the client
-type ValidationError error
